updated structure of order subsystem
