+++ .idea/modules.xml
@@ -2,8 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/Odosla 2.iml" filepath="$PROJECT_DIR$/.idea/Odosla 2.iml" />
-      <module fileurl="file://$PROJECT_DIR$/backend/backend.iml" filepath="$PROJECT_DIR$/backend/backend.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/Odosla.iml" filepath="$PROJECT_DIR$/.idea/Odosla.iml" />
     </modules>
   </component>
 </project>
