--- backend/src/main/java/order/OrderCreator.java
@@ -1,21 +1,0 @@
-package order;
-
-public class OrderCreator {
-    /* Attributes */
-    private Order order;
-    /* Constructor */
-    public OrderCreator(){
-
-    }
-    /* Getters and Setters */
-    public Order getOrder(){
-        return order;
-    }
-    public void addItem(Item item, int qty){
-
-    }
-    /* Functions */
-    public void clearOrder(){
-
-    }
-}
